feat: adding user id for authentication exception

+++ src/modules/tags/services/DeleteTag.ts
@@ -1,5 +1,6 @@
 import AppError from '@shared/errors/AppError';
 import { inject, injectable } from 'tsyringe';
+import { IGetOne } from '../domain/interfaces/IGetOne';
 import { ITagRepository } from '../domain/interfaces/ITagRepository';
 
 @injectable()
@@ -9,10 +10,14 @@ class DeleteTagService {
     private tagRepository: ITagRepository,
   ) {}
 
-  async execute(id: string): Promise<void> {
-    const hasTag = await this.tagRepository.findById(id);
+  async execute({ id, tag_id }: IGetOne): Promise<void> {
+    const hasTag = await this.tagRepository.findById(tag_id);
 
     if (!hasTag) throw new AppError('Tag not found', 404);
+
+    if (hasTag.user.id !== id) {
+      throw new AppError('Tag not found', 404);
+    }
 
     await this.tagRepository.deleteById(hasTag);
   }
